Major refactoring Moved various functions to separate files
