+++ .gitignore
@@ -173,3 +173,4 @@ dist
 
 output_data
 bin
+keyshares
